requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
Django==6.0.3
djangorestframework==3.17.1
djangorestframework-camel-case==1.4.2
idna==3.11
lxml==6.0.2
psycopg2-binary==2.9.11
pycparser==3.0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
sqlparse==0.5.5
urllib3==2.6.3
